# cocotola-question/api/openapi-bundled.yaml
openapi: 3.0.1
info:
  title: cocotola-question
  description: cocotola-question
  version: '1.0'
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers: []
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /api/v1/workbook:
    post:
      summary: Create a workbook
      deprecated: false
      description: Create a new workbook within the operator's space
      operationId: createWorkbook
      tags:
        - workbook
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkbookRequest'
            examples: {}
        required: true
      responses:
        '201':
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkbookResponse'
          headers: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
    get:
      summary: List workbooks
      deprecated: false
      description: List workbooks accessible by the authenticated user
      operationId: listWorkbooks
      tags:
        - workbook
      parameters: []
      responses:
        '200':
          description: Successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkbooksResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
  /api/v1/workbook/{workbookId}:
    get:
      summary: Get a workbook
      deprecated: false
      description: Get a specific workbook by ID
      operationId: getWorkbook
      tags:
        - workbook
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkbookResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Workbook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
    put:
      summary: Update a workbook
      deprecated: false
      description: Update an existing workbook
      operationId: updateWorkbook
      tags:
        - workbook
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkbookRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: Successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkbookResponse'
          headers: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Workbook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
    delete:
      summary: Delete a workbook
      deprecated: false
      description: Delete an existing workbook
      operationId: deleteWorkbook
      tags:
        - workbook
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully deleted
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Workbook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
...
components:
  schemas:
    CreateWorkbookRequest:
      type: object
      properties:
        spaceId:
          type: integer
          x-go-name: SpaceID
          format: int32
          x-oapi-codegen-extra-tags:
            binding: required
        title:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required,max=200
          maxLength: 200
          pattern: ^.*$
        description:
          type: string
          x-oapi-codegen-extra-tags:
            binding: max=1000
          maxLength: 1000
          pattern: ^.*$
        visibility:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required,oneof=private public
          enum:
            - private
            - public
      required:
        - spaceId
        - title
        - visibility
    UpdateWorkbookRequest:
      type: object
      properties:
        title:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required,max=200
          maxLength: 200
          pattern: ^.*$
        description:
          type: string
          x-oapi-codegen-extra-tags:
            binding: max=1000
          maxLength: 1000
          pattern: ^.*$
        visibility:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required,oneof=private public
          enum:
            - private
            - public
      required:
        - title
        - visibility
    WorkbookResponse:
      type: object
      properties:
        workbookId:
          type: string
          x-go-name: WorkbookID
          pattern: ^.*$
        spaceId:
          type: integer
          x-go-name: SpaceID
          format: int32
        ownerId:
          type: integer
          x-go-name: OwnerID
          format: int32
        organizationId:
          type: integer
          x-go-name: OrganizationID
          format: int32
        title:
          type: string
          pattern: ^.*$
        description:
          type: string
          pattern: ^.*$
        visibility:
          type: string
          enum:
            - private
            - public
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - workbookId
        - spaceId
        - ownerId
        - organizationId
        - title
        - description
        - visibility
        - createdAt
        - updatedAt
    ListWorkbooksResponse:
      type: object
      properties:
        workbooks:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/WorkbookResponse'
      required:
        - workbooks
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          pattern: ^.*$
        message:
          type: string
          pattern: ^.*$
      required:
        - code
        - message
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token